Transmogrifier beta, mteFunctions fixed
NPC Transmogrifier is in beta.  It's fully functional, but certain
aspects may still be modified before release.

mteFunctions' ElementByIP, which is the basis of several of other
function in the library, was compromised.  I'm not sure when this
happened or how it happened, but it has been fixed.

FMO text files are also getting pushed with this commit.  I'm not sure
why they weren't previously included.