FMO, mteFunctions, Break it Down, Re-Evaluator...
Updated lots of scripts.
FMO:
Almost ready for release.  See the changelog in its header.

mteFunctions:
Added HasItem and HasPerkCondition to mteFunctions, instead of having
them only in the Re-Evaluator.  They're pretty useful and it'll be nice
to have access to them in all my scripts.

Break it Down:
See the changelog in its header.