break it down update, FMO update
break it down has been fixed.  there was an error in a debug message in
the code and there stalhrim was misspelled.

FMO has a beta setup of installing the xp fix inside of the fmo patch.
I haven't gotten it to work yet, and have had trouble getting the script
to install the script files in the fmo\ folder.  I'll figure it out
eventually.