Armor Mod Builder update, Mator Smash Beta
Armor Mod Builder now has a button for selecting the meshes directory.
This is a lot more intuitive than typing a relative path manually.

Mator Smash is a simple script which allows for the creation of
rudimenty Bashed Patches.  It's currently not customizable and doesn't
have any tags, so conflicts are handled in a single way that maximizes
data.