Added Advanced Options window
separate advanced options window where all advanced options are now
available.  click to gear to get to it.
advanced options are saved to a config file when you click OK on the
advanced options window, and are loaded each time the script is run.
also have automatic detection of Mod Organizer's location