Numerous new features and fixes
Renumber only Conflicting FormIDs feature added.
Fixed Asset Copying to work with both MO and a normal "real" data
directory.
Fixed renumbering methods to not terminate prematurely, and made it so
injected records are skipped properly instead of getting renumbered.
Script will leave log window open after completion, though the current
implementation of this is not very good and needs to be fixed.