first run clarity, etc.
When a user first runs the script they'll be told to set up their
advanced options and will be taken directly to the Advanced Options
window.

Added a bunch of hints to the Advanced Options window which can be seen
on mouseover.  More to come.

Changed default renumbering mode to Renumber Conflicting FormIDs,
changed default value for usingMO to false.

If moPath is blank but usingMO is true, the script will terminate
instead of trying to perform a merge copying assets to a nonexistent
path.

Added a progress bar message for when the script finishes executing.
Progress window will automatically close if the log is not visible.

etc.