Various fixes
Fixing ignores for CopyAllAssets
Removing the "Removing..." messages for each record and instead using a
count for number of records removed.
Advanced Options window is more helpful and easier to use now.  Invalid
paths will not be accepted.
Added a verbose option to CopyDirectory.  Using Lowercase() for ignores
with CopyDirectory.