Merge Plugins, mteFunctions update
mteFunctions:
Lots of new functions.
- [DeleteDirectory]: deletes the contents of a directory, and optionally
the directory itself.
- [EditOutOfDate]: alerts the user that their xEdit is out of date, and
provides them with a button they can click to go to the AFKMods page to
download an updated version.
- [ConstructRadioGroup]: an all-in-one radiogroup constructor.
- [ConstructRadioButton]: an all-in-one radiobutton constructor.
- [ConstructEdit]: an all-in-one edit constructor.
- [cRadioGroup]: ConstructRadioGroup shortened function name.
- [cRadioButton]: ConstructRadioButton shortened function name.
- [cMemo]: ConstructMemo shortened function name.
- [cScrollBox]: ConstructScrollBox shortened function name.
- [cCheckBox]: ConstructCheckBox shortened function name.
- [cLabel]: ConstructLabel shortened function name.
- [cEdit]: ConstructEdit shortened function name.
- [cButton]: ConstructButton shortened function name.

Merge Plugins:
- Now using Edit Scripts\mp\temp as temporary directory, and verifying
files can be saved in the temporary directory.
- Fixed general asset copying failing due to exclude parameter issues
relating to spaces in the path to the temporary directory.  Now using a
local path to the exclude text document.
- Implemented user-friendly EditOutOfDate method from mteFunctions when
it has been detected that the user is using an out of date version of
xEdit.