fix GetDefinition, CopyGeneralAssets ignore
[Merge Plugins]
- Fixed issue with GetDefinition with v = true not working properly
because there's the letter "v" in the vs constant, but not in reports.
- Fixed issue with CopyGeneralAssets skipping files that should be
copied from facegendata.

[mteFunctions]
- Updated CopyDirectory and BatchCopyDirectory to not skip folders
matching ignore regular expressions containing '.' characters, except
'.' and '..', which are hardcoded to be ignored (as they were before
this update).  This is required for Merge Plugins v1.9 to work
correctly.  A possible better fix could be to have two different ignore
lists, one for files and one for folders (a la robocopy).