Mator Smash updates
This is v0.8.9.

I corrected an error from a refactoring in a previous commit
(BuildAllValuesList renamed to BuildElementList).

MergeUnsortedArray is now used instead of rcore for unsorted arrays.

dtArray element types are traversed with rcore, while etSubRecordArrays
are handled by MergeUnsortedArray if they are unsorted, or
MergeSortedArray if they are sorted.  NOTE: I don't know if dtArray
element types ever need to be merged with MergeUnsortedArray or
MergeSortedArray.  I've assumed they don't, but if we see issues with
arrays not merging properly this might be the cause.

New settings to show people how to use settings, and a modified default
setting to include unsorted arrays of elements (which should now be
handled by MergeUnsortedArray).